# refactor(verifier): Remove parallelism