--- a/Cargo.toml
+++ b/Cargo.toml
@@ -87,7 +87,7 @@ strum = { version = "0.26", features = ["derive"] }
 syn = "2.0"
 test-strategy = "0.4.0"
 thiserror = "2.0"
-twenty-first = "0.44.1"
+twenty-first = "0.45.0"
 unicode-width = "0.2"
 divan = "0.1.17"