--- a/triton-vm/src/proof_stream.rs
+++ b/triton-vm/src/proof_stream.rs
@@ -223,7 +223,7 @@ mod tests {
         let num_leaves = 1 << tree_height;
         let leaf_values: Vec<XFieldElement> = random_elements(num_leaves);
         let leaf_digests = leaf_values.iter().map(|&xfe| xfe.into()).collect_vec();
-        let merkle_tree = MerkleTree::new::<CpuParallel>(&leaf_digests).unwrap();
+        let merkle_tree = MerkleTree::par_new(&leaf_digests).unwrap();
         let indices_to_check = vec![5, 173, 175, 167, 228, 140, 252, 149, 232, 182, 5, 5, 182];
         let auth_structure = merkle_tree
             .authentication_structure(&indices_to_check)