--- a/triton-vm/src/shared_tests.rs
+++ b/triton-vm/src/shared_tests.rs
@@ -64,7 +64,7 @@ pub(crate) struct LeavedMerkleTreeTestData {
     #[strategy(Just(#leaves.iter().map(|&x| x.into()).collect()))]
     pub leaves_as_digests: Vec<Digest>,
 
-    #[strategy(Just(MerkleTree::new::<CpuParallel>(&#leaves_as_digests).unwrap()))]
+    #[strategy(Just(MerkleTree::par_new(&#leaves_as_digests).unwrap()))]
     pub merkle_tree: MerkleTree,
 
     #[strategy(Just(#revealed_indices.iter().map(|&i| #leaves[i]).collect()))]