@Sword-Smith 2025-01-08
This PR's branch, without parallelism:
```
Verify Halt             time:   [8.1461 ms 8.1681 ms 8.1947 ms]
```

`master`
```
time:   [11.644 ms 11.680 ms 11.735 ms] 
```

Removing parallelism makes verification faster, because of the overhead of rayon thread-handling.